- Fixed error during backup and restore. Now every backup is a clean backup. Snackbar is showing properly
