--- app/src/main/java/com/exercise/matipv2/data/repository/MatipRepository.kt
@@ -1,32 +1,0 @@
-package com.exercise.matipv2.data.repository
-
-import com.exercise.matipv2.data.local.model.List
-import com.exercise.matipv2.data.local.model.ListWithTips
-import com.exercise.matipv2.data.local.model.Tip
-import kotlinx.coroutines.flow.Flow
-
-interface MatipRepository {
-    /* Tip Methods */
-    suspend fun insertTip(tip: Tip)
-    suspend fun deleteTip(tip: Tip)
-    suspend fun updateTip(tip: Tip)
-    fun getAllTips(userId: String?): Flow<kotlin.collections.List<Tip>>
-    suspend fun getLastTipSaved(userId: String?): Tip
-    fun getAllTipsFromList(listId: Int, userId: String?): Flow<kotlin.collections.List<Tip>>
-    fun searchTipsInList(listId:  Int, query: String, userId: String?): Flow<kotlin.collections.List<Tip>>
-
-    /* List Methods */
-    suspend fun insertList(list: List)
-    suspend fun deleteList(list: List)
-    suspend fun updateList(list: List)
-    fun getAllLists(userId: String?): Flow<kotlin.collections.List<List>>
-    fun getListByName(listName: String, userId: String?): Flow<List>
-    fun getListById(listId: Int, userId: String?): Flow<List>
-    fun getAllListsWithTips(userId: String?): Flow<kotlin.collections.List<ListWithTips>>
-    fun searchLists(query:  String, userId: String?): Flow<kotlin.collections. List<List>>
-
-    suspend fun migrateGuestData(userId: String)
-    suspend fun backupDataToCloud(userId: String): Result<Unit>
-    suspend fun restoreDataFromCloud(userId: String): Result<Unit>
-    fun getLastBackupDate(userId: String): Flow<String?>
-}
