--- app/src/main/java/com/exercise/matipv2/data/repository/OfflineMatipRepository.kt
@@ -1,181 +1,0 @@
-package com.exercise.matipv2.data.repository
-
-import com.exercise.matipv2.data.local.dao.ListDao
-import com.exercise.matipv2.data.local.dao.TipDao
-import com.exercise.matipv2.data.local.model.List
-import com.exercise.matipv2.data.local.model.ListWithTips
-import com.exercise.matipv2.data.local.model.Tip
-import com.exercise.matipv2.util.localDateTimeFormated
-import com.google.firebase.firestore.FirebaseFirestore
-import com.google.firebase.firestore.SetOptions
-import kotlinx.coroutines.channels.awaitClose
-import kotlinx.coroutines.flow.Flow
-import kotlinx.coroutines.flow.callbackFlow
-import kotlinx.coroutines.flow.first
-import kotlinx.coroutines.tasks.await
-
-class OfflineMatipRepository(
-    private val tipDao: TipDao,
-    private val listDao: ListDao,
-    private val firestore: FirebaseFirestore
-) : MatipRepository {
-
-    private fun userRef(userId: String) = firestore.collection("users").document(userId)
-    private fun userListsRef(userId: String) = userRef(userId).collection("lists")
-    private fun userTipsRef(userId: String) = userRef(userId).collection("tips")
-
-    // Tip Methods
-    override suspend fun insertTip(tip: Tip) {
-        tipDao.insertTip(tip)
-    }
-
-    override suspend fun deleteTip(tip: Tip) {
-        tipDao.deleteTip(tip)
-    }
-
-    override suspend fun updateTip(tip: Tip) {
-        tipDao.updateTip(tip)
-    }
-
-    override fun getAllTips(userId: String?) = tipDao.getAllTips(userId)
-    
-    override suspend fun getLastTipSaved(userId: String?): Tip {
-        return tipDao.getLastTipSaved(userId)
-    }
-
-    override fun getAllTipsFromList(listId: Int, userId: String?): Flow<kotlin.collections.List<Tip>> {
-        return tipDao.getAllTipsFromList(listId, userId)
-    }
-
-    override fun searchTipsInList(listId: Int, query: String, userId: String?) = 
-        tipDao.searchTipsInList(listId, query, userId)
-
-
-    // List Methods
-    override suspend fun insertList(list: List) {
-        listDao.insertList(list)
-    }
-
-    override suspend fun deleteList(list: List) {
-        listDao.deleteList(list)
-    }
-
-    override suspend fun updateList(list: List) {
-        listDao.updateList(list)
-    }
-
-    override fun getAllLists(userId: String?) = listDao.getAllLists(userId)
-    
-    override fun getAllListsWithTips(userId: String?): Flow<kotlin.collections.List<ListWithTips>> {
-        return listDao.getAllListsWithTips(userId)
-    }
-
-    override fun getListByName(listName: String, userId: String?): Flow<List> {
-        return listDao.getListByName(listName, userId)
-    }
-
-    override fun getListById(listId: Int, userId: String?): Flow<List> {
-        return listDao.getListById(listId, userId)
-    }
-
-    override fun searchLists(query: String, userId: String?) = listDao.searchLists(query, userId)
-
-    override suspend fun migrateGuestData(userId: String) {
-        listDao.migrateGuestLists(userId)
-        tipDao.migrateGuestTips(userId)
-    }
-
-    override suspend fun backupDataToCloud(userId: String): Result<Unit> {
-        return try {
-            val batch = firestore.batch()
-            
-            // 1. Get current local data state
-            val localLists = listDao.getAllLists(userId).first()
-            val localTips = tipDao.getAllTips(userId).first()
-            
-            val localListIds = localLists.map { it.id.toString() }.toSet()
-            val localTipIds = localTips.map { it.id.toString() }.toSet()
-
-            // 2. Clear removed items from Cloud
-            val cloudLists = userListsRef(userId).get().await()
-            for (doc in cloudLists.documents) {
-                if (doc.id !in localListIds) {
-                    batch.delete(doc.reference)
-                }
-            }
-            
-            val cloudTips = userTipsRef(userId).get().await()
-            for (doc in cloudTips.documents) {
-                if (doc.id !in localTipIds) {
-                    batch.delete(doc.reference)
-                }
-            }
-
-            // 3. Upload/Update current items
-            for (list in localLists) {
-                val docRef = userListsRef(userId).document(list.id.toString())
-                batch.set(docRef, list)
-            }
-            for (tip in localTips) {
-                val docRef = userTipsRef(userId).document(tip.id.toString())
-                batch.set(docRef, tip)
-            }
-            
-            // 4. Update last backup timestamp
-            val timestamp = localDateTimeFormated()
-            batch.set(userRef(userId), mapOf("last_backup" to timestamp), SetOptions.merge())
-            
-            batch.commit().await()
-            Result.success(Unit)
-        } catch (e: Exception) {
-            Result.failure(e)
-        }
-    }
-
-    override suspend fun restoreDataFromCloud(userId: String): Result<Unit> {
-        return try {
-            val cloudLists = userListsRef(userId).get().await()
-            val cloudTips = userTipsRef(userId).get().await()
-
-            if (cloudLists.isEmpty && cloudTips.isEmpty) {
-                return Result.failure(Exception("No backup found"))
-            }
-
-            // 1. Clean current local user state to prevent duplicates/summing
-            tipDao.deleteAllTipsForUser(userId)
-            listDao.deleteAllListsForUser(userId)
-
-            // 2. Insert cloud snapshot
-            for (doc in cloudLists.documents) {
-                val list = doc.toObject(List::class.java)
-                if (list != null) {
-                    // Important: Keep original ID if possible, or reset to 0 for autogen.
-                    // To maintain List-Tip relationship, we should keep original IDs.
-                    listDao.insertList(list) 
-                }
-            }
-            for (doc in cloudTips.documents) {
-                val tip = doc.toObject(Tip::class.java)
-                if (tip != null) {
-                    tipDao.insertTip(tip)
-                }
-            }
-            
-            Result.success(Unit)
-        } catch (e: Exception) {
-            Result.failure(e)
-        }
-    }
-
-    override fun getLastBackupDate(userId: String): Flow<String?> = callbackFlow {
-        val listener = userRef(userId).addSnapshotListener { snapshot, error ->
-            if (error != null) {
-                trySend(null)
-                return@addSnapshotListener
-            }
-            val date = snapshot?.getString("last_backup")
-            trySend(date)
-        }
-        awaitClose { listener.remove() }
-    }
-}
