+++ app/src/main/java/com/exercise/matipv2/di/AppModule.kt
@@ -3,9 +3,11 @@ package com.exercise.matipv2.di
 import com.exercise.matipv2.data.analytics.AnalyticsHelper
 import com.exercise.matipv2.data.analytics.FirebaseAnalyticsHelper
 import com.exercise.matipv2.data.repository.AuthRepository
+import com.exercise.matipv2.data.repository.BackupRepository
 import com.exercise.matipv2.data.repository.FirebaseAuthRepository
-import com.exercise.matipv2.data.repository.MatipRepository
-import com.exercise.matipv2.data.repository.OfflineMatipRepository
+import com.exercise.matipv2.data.repository.FirestoreBackupRepository
+import com.exercise.matipv2.data.repository.LocalRepository
+import com.exercise.matipv2.data.repository.OfflineLocalRepository
 import com.exercise.matipv2.ui.MainScreenViewModel
 import com.exercise.matipv2.ui.auth.AuthViewModel
 import com.google.firebase.firestore.FirebaseFirestore
@@ -16,7 +18,8 @@ import org.koin.core.module.dsl.viewModelOf
 import org.koin.dsl.module
 
 val appModule = module {
-    singleOf(::OfflineMatipRepository) { bind<MatipRepository>() }
+    singleOf(::OfflineLocalRepository) { bind<LocalRepository>() }
+    singleOf(::FirestoreBackupRepository) { bind<BackupRepository>() }
     single<AuthRepository> { FirebaseAuthRepository(androidContext()) }
     single<AnalyticsHelper> { FirebaseAnalyticsHelper(androidContext()) }
     single { FirebaseFirestore.getInstance() }
